rename and add changeset file
