+++ .changeset/infuie-feir-wmqw.md
@@ -1,0 +1,5 @@
+---
+'@ldn-viz/ui': minor
+---
+
+ADDED - add a `MergeValuesControl` that allows users to define how different values of a categorical variable should be merged into new categories
